softgpu: Move clipping code to a separate file.
