+++ CMakeLists.txt
@@ -1015,12 +1015,14 @@ add_library(GPU OBJECT
 	GPU/Math3D.h
 	GPU/Null/NullGpu.cpp
 	GPU/Null/NullGpu.h
+	GPU/Software/Clipper.cpp
+	GPU/Software/Clipper.h
+	GPU/Software/Rasterizer.cpp
+	GPU/Software/Rasterizer.h
 	GPU/Software/SoftGpu.cpp
 	GPU/Software/SoftGpu.h
 	GPU/Software/TransformUnit.cpp
 	GPU/Software/TransformUnit.h
-	GPU/Software/Rasterizer.cpp
-	GPU/Software/Rasterizer.h
 	GPU/ge_constants.h)
 setup_target_project(GPU GPU)
